+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/
 ## Unreleased
 - Fixed iOS 17 BUtton issue.
 - Fixed the Typing Customization issue.
+- Added hidepostCTA support for all types of buttons.
 
 ## [1.2.0] 2023-10-27
 - Fixed Button Spacing