+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 ## Unreleased
+- Fixed buttons are getting cut in form template
 - Added hidepostCTA support for all types of buttons.
 
 ## [1.2.0] 2023-10-27