+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
 ## Unreleased
+- Fixed iOS 17 BUtton issue.
+- Fixed the Typing Customization issue.
 - Fixed buttons are getting cut in form template
 - Added hidepostCTA support for all types of buttons.