+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
-## Unreleased
+## [1.2.1] 2023-11-02
 - Fixed iOS 17 BUtton issue.
 - Fixed the Typing Customization issue.
 - Fixed buttons are getting cut in form template